+++ tsdb/engine/tsm1/engine.go
@@ -23,6 +23,7 @@ import (
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/pkg/bytesutil"
 	"github.com/influxdata/influxdb/pkg/estimator"
+	"github.com/influxdata/influxdb/pkg/limiter"
 	"github.com/influxdata/influxdb/tsdb"
 	_ "github.com/influxdata/influxdb/tsdb/index"
 	"github.com/uber-go/zap"
@@ -132,6 +133,9 @@ type Engine struct {
 	enableCompactionsOnOpen bool
 
 	stats *EngineStatistics
+
+	// The limiter for concurrent compactions
+	compactionLimiter limiter.Fixed
 }
 
 // NewEngine returns a new instance of Engine.
@@ -161,17 +165,15 @@ func NewEngine(id uint64, idx tsdb.Index, path string, walPath string, opt tsdb.
 		WAL:   w,
 		Cache: cache,
 
-		FileStore: fs,
-		Compactor: c,
-		CompactionPlan: &DefaultPlanner{
-			FileStore:                    fs,
-			CompactFullWriteColdDuration: time.Duration(opt.Config.CompactFullWriteColdDuration),
-		},
+		FileStore:      fs,
+		Compactor:      c,
+		CompactionPlan: NewDefaultPlanner(fs, time.Duration(opt.Config.CompactFullWriteColdDuration)),
 
 		CacheFlushMemorySizeThreshold: opt.Config.CacheSnapshotMemorySize,
 		CacheFlushWriteColdDuration:   time.Duration(opt.Config.CacheSnapshotWriteColdDuration),
 		enableCompactionsOnOpen:       true,
-		stats: &EngineStatistics{},
+		stats:             &EngineStatistics{},
+		compactionLimiter: opt.CompactionLimiter,
 	}
 
 	// Attach fieldset to index.
@@ -243,6 +245,7 @@ func (e *Engine) disableLevelCompactions(wait bool) {
 		e.levelWorkers += 1
 	}
 
+	var cleanup bool
 	if old == 0 && e.done != nil {
 		// Prevent new compactions from starting
 		e.Compactor.DisableCompactions()
@@ -250,12 +253,13 @@ func (e *Engine) disableLevelCompactions(wait bool) {
 		// Stop all background compaction goroutines
 		close(e.done)
 		e.done = nil
+		cleanup = true
 	}
 
 	e.mu.Unlock()
 	e.wg.Wait()
 
-	if old == 0 { // first to disable should cleanup
+	if cleanup { // first to disable should cleanup
 		if err := e.cleanup(); err != nil {
 			e.logger.Info(fmt.Sprintf("error cleaning up temp file: %v", err))
 		}
@@ -428,6 +432,11 @@ func (e *Engine) Statistics(tags map[string]string) []models.Statistic {
 	return statistics
 }
 
+// DiskSize returns the total size in bytes of all TSM and WAL segments on disk.
+func (e *Engine) DiskSize() int64 {
+	return e.FileStore.DiskSizeBytes() + e.WAL.DiskSizeBytes()
+}
+
 // Open opens and initializes the engine.
 func (e *Engine) Open() error {
 	if err := os.MkdirAll(e.path, 0777); err != nil {
@@ -524,6 +533,21 @@ func (e *Engine) LoadMetadataIndex(shardID uint64, index tsdb.Index) error {
 
 	e.traceLogger.Info(fmt.Sprintf("Meta data index for shard %d loaded in %v", shardID, time.Since(now)))
 	return nil
+}
+
+// IsIdle returns true if the cache is empty, there are no running compactions and the
+// shard is fully compacted.
+func (e *Engine) IsIdle() bool {
+	cacheEmpty := e.Cache.Size() == 0
+
+	runningCompactions := atomic.LoadInt64(&e.stats.CacheCompactionsActive)
+	runningCompactions += atomic.LoadInt64(&e.stats.TSMCompactionsActive[0])
+	runningCompactions += atomic.LoadInt64(&e.stats.TSMCompactionsActive[1])
+	runningCompactions += atomic.LoadInt64(&e.stats.TSMCompactionsActive[2])
+	runningCompactions += atomic.LoadInt64(&e.stats.TSMFullCompactionsActive)
+	runningCompactions += atomic.LoadInt64(&e.stats.TSMOptimizeCompactionsActive)
+
+	return cacheEmpty && runningCompactions == 0 && e.CompactionPlan.FullyCompacted()
 }
 
 // Backup writes a tar archive of any TSM files modified since the passed
@@ -1165,8 +1189,12 @@ func (e *Engine) compactTSMLevel(fast bool, level int, quit <-chan struct{}) {
 		case <-t.C:
 			s := e.levelCompactionStrategy(fast, level)
 			if s != nil {
+				// Release the files in the compaction plan
+				defer e.CompactionPlan.Release(s.compactionGroups)
+
 				s.Apply()
 			}
+
 		}
 	}
 }
@@ -1183,6 +1211,8 @@ func (e *Engine) compactTSMFull(quit <-chan struct{}) {
 		case <-t.C:
 			s := e.fullCompactionStrategy()
 			if s != nil {
+				// Release the files in the compaction plan
+				defer e.CompactionPlan.Release(s.compactionGroups)
 				s.Apply()
 			}
 
@@ -1205,6 +1235,7 @@ type compactionStrategy struct {
 	logger    zap.Logger
 	compactor *Compactor
 	fileStore *FileStore
+	limiter   limiter.Fixed
 }
 
 // Apply concurrently compacts all the groups in a compaction strategy.
@@ -1226,6 +1257,12 @@ func (s *compactionStrategy) Apply() {
 
 // compactGroup executes the compaction strategy against a single CompactionGroup.
 func (s *compactionStrategy) compactGroup(groupNum int) {
+	// Limit concurrent compactions if we have a limiter
+	if cap(s.limiter) > 0 {
+		s.limiter.Take()
+		defer s.limiter.Release()
+	}
+
 	group := s.compactionGroups[groupNum]
 	start := time.Now()
 	s.logger.Info(fmt.Sprintf("beginning %s compaction of group %d, %d TSM files", s.description, groupNum, len(group)))
@@ -1290,6 +1327,7 @@ func (e *Engine) levelCompactionStrategy(fast bool, level int) *compactionStrate
 		fileStore:        e.FileStore,
 		compactor:        e.Compactor,
 		fast:             fast,
+		limiter:          e.compactionLimiter,
 
 		description:  fmt.Sprintf("level %d", level),
 		activeStat:   &e.stats.TSMCompactionsActive[level-1],
@@ -1320,6 +1358,7 @@ func (e *Engine) fullCompactionStrategy() *compactionStrategy {
 		fileStore:        e.FileStore,
 		compactor:        e.Compactor,
 		fast:             optimize,
+		limiter:          e.compactionLimiter,
 	}
 
 	if optimize {