+++ tsdb/shard.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"math"
-	"os"
 	"path/filepath"
 	"regexp"
 	"sort"
@@ -110,6 +109,7 @@ type Shard struct {
 	path    string
 	walPath string
 	id      uint64
+	wg      sync.WaitGroup
 
 	database        string
 	retentionPolicy string
@@ -206,14 +206,9 @@ func (s *Shard) Statistics(tags map[string]string) []models.Statistic {
 		return nil
 	}
 
-	// TODO(edd): Should statSeriesCreate be the current number of series in the
-	// shard, or the total number of series ever created?
-	sSketch, tSketch, err := s.engine.SeriesSketches()
-	seriesN := int64(sSketch.Count() - tSketch.Count())
-	if err != nil {
-		s.logger.Error("cannot compute series sketch", zap.Error(err))
-		seriesN = 0
-	}
+	// Refresh our disk size stat
+	_, _ = s.DiskSize()
+	seriesN := s.engine.SeriesN()
 
 	tags = s.defaultTags.Merge(tags)
 	statistics := []models.Statistic{{
@@ -288,8 +283,6 @@ func (s *Shard) Open() error {
 		}
 		s.engine = e
 
-		go s.monitor()
-
 		return nil
 	}(); err != nil {
 		s.close(true)
@@ -335,6 +328,7 @@ func (s *Shard) close(clean bool) error {
 	default:
 		close(s.closing)
 	}
+	s.wg.Wait()
 
 	if clean {
 		// Don't leak our shard ID and series keys in the index
@@ -350,6 +344,12 @@ func (s *Shard) close(clean bool) error {
 		s.index = nil
 	}
 	return err
+}
+
+func (s *Shard) IndexType() string {
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+	return s.index.Type()
 }
 
 // ready determines if the Shard is ready for queries or writes.
@@ -380,35 +380,28 @@ func (s *Shard) UnloadIndex() {
 	s.index.RemoveShard(s.id)
 }
 
-// DiskSize returns the size on disk of this shard
-func (s *Shard) DiskSize() (int64, error) {
-	var size int64
-	err := filepath.Walk(s.path, func(_ string, fi os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if !fi.IsDir() {
-			size += fi.Size()
-		}
-		return err
-	})
-	if err != nil {
-		return 0, err
+// IsIdle return true if the shard is not receiving writes and is fully compacted.
+func (s *Shard) IsIdle() bool {
+	if err := s.ready(); err != nil {
+		return true
 	}
 
-	err = filepath.Walk(s.walPath, func(_ string, fi os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
+	return s.engine.IsIdle()
+}
 
-		if !fi.IsDir() {
-			size += fi.Size()
-		}
-		return err
-	})
+// SetCompactionsEnabled enables or disable shard background compactions.
+func (s *Shard) SetCompactionsEnabled(enabled bool) {
+	if err := s.ready(); err != nil {
+		return
+	}
+	s.engine.SetCompactionsEnabled(enabled)
+}
 
-	return size, err
+// DiskSize returns the size on disk of this shard
+func (s *Shard) DiskSize() (int64, error) {
+	size := s.engine.DiskSize()
+	atomic.StoreInt64(&s.stats.DiskBytes, size)
+	return size, nil
 }
 
 // FieldCreate holds information for a field to create on a measurement.
@@ -964,62 +957,12 @@ func (s *Shard) CreateSnapshot() (string, error) {
 	return s.engine.CreateSnapshot()
 }
 
-func (s *Shard) monitor() {
-	t := time.NewTicker(monitorStatInterval)
-	defer t.Stop()
-	t2 := time.NewTicker(time.Minute)
-	defer t2.Stop()
-	var changed time.Time
+func (s *Shard) ForEachMeasurementTagKey(name []byte, fn func(key []byte) error) error {
+	return s.engine.ForEachMeasurementTagKey(name, fn)
+}
 
-	for {
-		select {
-		case <-s.closing:
-			return
-		case <-t.C:
-
-			// Checking DiskSize can be expensive with a lot of shards and TSM files, only
-			// check if something has changed.
-			lm := s.LastModified()
-			if lm.Equal(changed) {
-				continue
-			}
-
-			size, err := s.DiskSize()
-			if err != nil {
-				s.logger.Info(fmt.Sprintf("Error collecting shard size: %v", err))
-				continue
-			}
-			atomic.StoreInt64(&s.stats.DiskBytes, size)
-			changed = lm
-		case <-t2.C:
-			if s.options.Config.MaxValuesPerTag == 0 {
-				continue
-			}
-
-			names, err := s.MeasurementNamesByExpr(nil)
-			if err != nil {
-				s.logger.Warn("cannot retrieve measurement names", zap.Error(err))
-				continue
-			}
-
-			for _, name := range names {
-				s.engine.ForEachMeasurementTagKey(name, func(k []byte) error {
-					n := s.engine.TagKeyCardinality(name, k)
-					perc := int(float64(n) / float64(s.options.Config.MaxValuesPerTag) * 100)
-					if perc > 100 {
-						perc = 100
-					}
-
-					// Log at 80, 85, 90-100% levels
-					if perc == 80 || perc == 85 || perc >= 90 {
-						s.logger.Info(fmt.Sprintf("WARN: %d%% of max-values-per-tag limit exceeded: (%d/%d), db=%s shard=%d measurement=%s tag=%s",
-							perc, n, s.options.Config.MaxValuesPerTag, s.database, s.id, name, k))
-					}
-					return nil
-				})
-			}
-		}
-	}
+func (s *Shard) TagKeyCardinality(name, key []byte) int {
+	return s.engine.TagKeyCardinality(name, key)
 }
 
 type ShardGroup interface {